bug fix
the bug we fixed caused a list index out of range error for the self.loopp list. this was due to the fact that the on_draw() method ran before the ThreadedNetwork thread, which made it so we couldn't add a player index to the self.loopp list. The fix was simply a time.sleep() at the init function